% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lunarCal.R
\name{as.lunar}
\alias{as.lunar}
\title{Lunar date conversion function from character}
\usage{
as.lunar(x, check = TRUE)
}
\arguments{
\item{x}{String}
}
\value{
Lunar Date
}
\description{
A handy way to construct lunar date, akin the as.Date().
}

