% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lunarCal.R
\name{formatLunar}
\alias{formatLunar}
\title{Format lunar date into Chinese string}
\usage{
formatLunar(lunarDate, withZodiac = FALSE)
}
\arguments{
\item{lunarDate}{date to convert}

\item{withZodiac}{Append the Chinese Zodiac sign to the string output of
lunar date. Using Cantonese version of Chinese Zodiac signs.}
}
\value{
Traditional Chinese string representation of the lunar date
}
\description{
convert lunar date to Traditional Chinese representation.
}

