% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lunarCal.R
\name{is.lunar}
\alias{is.lunar}
\title{Check for the validity of lunar date}
\usage{
is.lunar(lunarDate)
}
\arguments{
\item{lunarDate}{date to check}
}
\value{
Boolean
}
\description{
Check the validty of lunar date
}

