\name{formatLunar}
\alias{formatLunar}
\title{Format lunar date into Chinese string}
\usage{
  formatLunar(lunarDate, withZodiac = FALSE)
}
\arguments{
  \item{lunarDate}{date to convert}

  \item{withZodiac}{Append the Chinese Zodiac sign to the
  string output of lunar date}
}
\value{
  A Traditional Chinese string representation of the lunar
  date
}
\description{
  This function converts lunar date to Chinese string.
}

