% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lunarCal.R
\name{lunarCal}
\alias{lunarCal}
\title{Convert Solar date to and from lunar date}
\usage{
lunarCal(x, toString = FALSE, withZodiac = FALSE, ignoreLeap = TRUE)
}
\arguments{
\item{x}{date to convert, can be solar date or lunar date. Date object is converted to lunar date. Vector with
the format of lunar date is converted to solar date.}

\item{toString}{format the output lunar date to Chinese string (Traditional Chinese)}

\item{withZodiac}{Append the Chinese Zodiac sign to the string output of
lunar date}

\item{ignoreLeap}{ignore leap month when the converted lunar date can have
a leap month}
}
\value{
Date object (solar date) or lunar date, depends on the input x.
}
\description{
This function allows the conversion between solar date and lunar date.
}

